analytics_engine/requirements.txt
flask==3.0.0
pandas==2.1.4
scikit-learn==1.3.2
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
